# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
  domain.LoginRequest:
    properties:
      password:
        maxLength: 100
        minLength: 6
        type: string
      username:
        description: can be email or phone
        maxLength: 255
        minLength: 3
        type: string
    required:
    - password
    - username
    type: object
  domain.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  domain.RegisterRequest:
    properties:
      email:
        maxLength: 255
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      password:
        description: Stronger minimum
        maxLength: 100
        minLength: 8
        type: string
      phone:
        description: E.164 format
        maxLength: 15
        minLength: 10
        type: string
    required:
    - email
    - name
    - password
    - phone
    type: object
  dto.CreateInvoiceRequest:
    properties:
      customer:
        maxLength: 200
        minLength: 1
        type: string
      issue_date:
        type: string
      issuer:
        maxLength: 200
        minLength: 1
        type: string
      items:
        items:
          $ref: '#/definitions/dto.InvoiceItemRequest'
        minItems: 1
        type: array
      note:
        maxLength: 1000
        type: string
    required:
    - customer
    - issue_date
    - issuer
    - items
    type: object
  dto.InvoiceItemRequest:
    properties:
      description:
        maxLength: 500
        minLength: 1
        type: string
      price:
        minimum: 0
        type: integer
      qty:
        minimum: 1
        type: integer
    required:
    - description
    - price
    - qty
    type: object
  http.Resp:
    properties:
      data: {}
      errors:
        items:
          type: string
        type: array
      meta: {}
      request_id:
        type: string
      title:
        type: string
      version:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Xonvera Service API - Cashflow and Invoice Management
  title: Xonvera API
  version: 1.0.0
paths:
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Login with email or phone and password
      parameters:
      - description: Login Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.Resp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.Resp'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.Resp'
      summary: Login user
      tags:
      - Auth
  /api/v1/auth/logout:
    post:
      consumes:
      - application/json
      description: Invalidate current access token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.Resp'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.Resp'
      security:
      - BearerAuth: []
      summary: Logout user
      tags:
      - Auth
  /api/v1/auth/refresh:
    post:
      consumes:
      - application/json
      description: Get new access token using refresh token
      parameters:
      - description: Refresh Token Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.Resp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.Resp'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.Resp'
      summary: Refresh access token
      tags:
      - Auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user with name, email, phone, and password
      parameters:
      - description: Register Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.Resp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.Resp'
      summary: Register a new user
      tags:
      - Auth
  /api/v1/invoices:
    get:
      consumes:
      - application/json
      description: Get all invoices with pagination
      parameters:
      - default: 20
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.Resp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.Resp'
      summary: Get all invoices
      tags:
      - Invoice
    post:
      consumes:
      - application/json
      description: Create a new invoice with items
      parameters:
      - description: Create Invoice Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateInvoiceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.Resp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.Resp'
      summary: Create a new invoice
      tags:
      - Invoice
  /api/v1/invoices/{id}:
    get:
      consumes:
      - application/json
      description: Get invoice details by ID
      parameters:
      - description: Invoice ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.Resp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.Resp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.Resp'
      summary: Get invoice by ID
      tags:
      - Invoice
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"